update: move general fucntions to utils
